+++ doc/history.rst
@@ -14,6 +14,8 @@ Version 3.3
 * Removed the ``pexpect.psh`` module. This was never documented, and we found
   no evidence that people use it. The new :mod:`pexpect.replwrap` module
   provides a more flexible alternative.
+* Fixed ``TypeError: got <type 'str'> ('\r\n') as pattern`` in ``readline()``
+  method of ``spawnu`` (:ghissue:`67`).
 
 Version 3.2
 ```````````